move rand package to core
