+++ core/rand/text.go
